paper_agent_cong/requirements.txt
PyYAML>=6.0
pypdf>=5.0.0
pyzotero>=1.5.8
